Data Augmentation and Synthetic/requirements.txt
google-generativeai
PyPDF2
numpy
pandas
tqdm